refactor: 평수 입력시 소수점 입력 가능하게 변경

+++ src/main/java/com/theroom/server/domain/entity/Contact.java
@@ -45,7 +45,7 @@ public class Contact {
     private String customer;
     private String email;
     private String phoneNumber;
-    private int exclusiveArea;
+    private double exclusiveArea;
     private boolean personalInformationAgree;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
