+++ src/main/java/com/theroom/server/domain/request/ContactAddRequest.java
@@ -19,7 +19,7 @@ public class ContactAddRequest {
     private String email;
     private String phoneNumber;
     private BuildingType buildingType;
-    private int exclusiveArea;
+    private double exclusiveArea;
     private double latitude;
     private double longitude;
     private String postCode;
